Moved packages from Store reducer to Delivery reducer
