+++ src/redux/Delivery/selectors.js
@@ -1,3 +1,4 @@
+export const selectPackages = state => state.delivery.packages;
 export const selectStore = state => state.delivery.store;
-export const selectTimeSlots = state => state.delivery.timeSlots;
 export const selectTimeSlotChoices = state => state.delivery.timeSlotChoices;
+export const selectTimeSlots = state => state.delivery.timeSlots;
